requirements.txt
asgiref==3.3.1
Django==2.2
Jinja2==2.11.3
MarkupSafe==1.1.1
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.12.2
wheel==0.36.2
psycopg2==2.8.6  # Requires python3-dev
pytest==6.2.2
pytest-bdd==4.0.2
requests==2.25.1
